moved svg to images
