--- src/components/SettingModal/SVGInput.jsx
@@ -1,21 +1,0 @@
-const SVGInput = (props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={134}
-    height={18}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="#407BFF"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M2 12v1.5A1.5 1.5 0 0 0 3.5 15h9a1.5 1.5 0 0 0 1.5-1.5V12M5 6l3-3m0 0 3 3M8 3v9"
-    />
-    <path
-      fill="#407BFF"
-      d="M30.597 4.047h1.709v6.651c0 .757-.164 1.39-.492 1.9-.329.511-.775.896-1.34 1.156-.56.255-1.188.383-1.88.383-.716 0-1.354-.128-1.914-.383a3.057 3.057 0 0 1-1.326-1.155c-.32-.51-.479-1.144-.479-1.9V4.046h1.709v6.651c0 .479.082.873.246 1.183.164.305.397.53.697.677.301.146.657.218 1.067.218.41 0 .763-.072 1.06-.218.3-.146.532-.372.697-.677.164-.31.246-.704.246-1.183V4.047Zm5.065 3.978v8.819h-1.647V6.604h1.517l.13 1.421Zm4.82 2.208v.144c0 .538-.064 1.037-.192 1.497a3.656 3.656 0 0 1-.554 1.196c-.241.338-.54.6-.895.786-.356.187-.766.28-1.23.28-.46 0-.864-.084-1.21-.252a2.48 2.48 0 0 1-.869-.732 3.969 3.969 0 0 1-.574-1.107 7.113 7.113 0 0 1-.3-1.408v-.554c.059-.542.159-1.034.3-1.477a4 4 0 0 1 .574-1.141c.237-.32.527-.565.868-.738.342-.174.741-.26 1.197-.26.465 0 .877.09 1.237.273.36.178.663.433.91.766.245.328.43.724.553 1.19.123.46.184.972.184 1.537Zm-1.648.144v-.144c0-.341-.032-.658-.096-.95a2.482 2.482 0 0 0-.3-.78 1.473 1.473 0 0 0-.527-.519 1.432 1.432 0 0 0-.759-.191c-.291 0-.542.05-.752.15-.21.096-.385.23-.526.404-.141.173-.25.376-.328.608a3.644 3.644 0 0 0-.164.745v1.326c.055.329.148.63.28.903.132.273.32.492.56.656.247.16.561.24.944.24.296 0 .55-.065.759-.192.21-.128.38-.303.513-.527.136-.227.236-.49.3-.786.064-.296.096-.61.096-.943ZM43.647 3.5V14h-1.655V3.5h1.655Zm1.49 6.884v-.157c0-.534.077-1.028.232-1.484.155-.46.378-.859.67-1.196a3.023 3.023 0 0 1 1.08-.793 3.516 3.516 0 0 1 1.45-.287c.542 0 1.025.096 1.449.287.428.187.79.451 1.087.793.296.337.521.736.676 1.196.155.456.233.95.233 1.484v.157c0 .533-.078 1.027-.233 1.483a3.55 3.55 0 0 1-.676 1.196 3.136 3.136 0 0 1-1.08.793 3.54 3.54 0 0 1-1.443.28 3.607 3.607 0 0 1-1.456-.28 3.135 3.135 0 0 1-1.08-.793 3.551 3.551 0 0 1-.677-1.196 4.582 4.582 0 0 1-.232-1.483Zm1.647-.157v.157c0 .332.034.647.103.943a2.4 2.4 0 0 0 .321.78c.146.223.333.398.56.526.228.127.5.191.814.191.305 0 .57-.064.793-.191.228-.128.415-.303.56-.527.146-.223.253-.483.322-.779.073-.296.11-.61.11-.943v-.157c0-.329-.037-.638-.11-.93a2.387 2.387 0 0 0-.328-.786 1.569 1.569 0 0 0-.56-.533 1.542 1.542 0 0 0-.8-.199c-.31 0-.58.066-.807.199a1.587 1.587 0 0 0-.554.533c-.146.228-.253.49-.321.786a4.068 4.068 0 0 0-.103.93Zm10.78 2.29V8.989c0-.264-.047-.492-.143-.683a1.005 1.005 0 0 0-.438-.445c-.191-.104-.432-.157-.724-.157a1.64 1.64 0 0 0-.697.137 1.128 1.128 0 0 0-.459.369.89.89 0 0 0-.163.526h-1.641c0-.291.07-.574.212-.847.141-.274.346-.518.615-.732.269-.214.59-.383.964-.506a4.01 4.01 0 0 1 1.258-.184c.556 0 1.048.093 1.476.28.433.187.773.47 1.019.848.25.373.376.843.376 1.408v3.288c0 .337.023.64.068.91.05.263.12.494.212.69V14H57.81a2.918 2.918 0 0 1-.184-.677 5.528 5.528 0 0 1-.061-.806Zm.24-3.015.013 1.018h-1.182c-.306 0-.575.03-.807.09a1.656 1.656 0 0 0-.58.245c-.156.11-.272.242-.35.397-.077.155-.116.33-.116.526s.046.376.137.54c.091.16.223.285.396.376.178.091.392.137.643.137.337 0 .631-.068.882-.205.255-.141.456-.312.601-.513.146-.205.224-.399.233-.58l.533.73a2.506 2.506 0 0 1-.28.602c-.132.214-.306.42-.52.615a2.584 2.584 0 0 1-1.77.656c-.483 0-.914-.095-1.292-.286a2.297 2.297 0 0 1-.889-.787 2.039 2.039 0 0 1-.321-1.128c0-.391.073-.738.218-1.039.15-.305.37-.56.657-.765.291-.205.647-.36 1.066-.465.42-.11.898-.164 1.436-.164h1.292Zm7.65 2.967V3.5h1.653V14H65.61l-.157-1.531Zm-4.813-2.085v-.144c0-.56.066-1.07.198-1.531.132-.465.323-.864.574-1.196.25-.338.556-.595.916-.773.36-.182.766-.273 1.217-.273.447 0 .838.086 1.176.26.337.173.624.421.861.745.237.319.426.702.567 1.148.142.442.242.934.301 1.477v.458a7.4 7.4 0 0 1-.3 1.449c-.142.437-.331.816-.568 1.135a2.453 2.453 0 0 1-.868.738c-.337.173-.732.26-1.183.26-.446 0-.85-.094-1.21-.28a2.673 2.673 0 0 1-.909-.787 3.699 3.699 0 0 1-.574-1.19 5.405 5.405 0 0 1-.198-1.496Zm1.647-.144v.144c0 .337.03.652.089.943.064.292.162.55.294.773.132.218.303.392.513.52.214.122.469.184.765.184.374 0 .681-.082.923-.246a1.68 1.68 0 0 0 .567-.663c.142-.283.237-.598.288-.944V9.714a3.023 3.023 0 0 0-.171-.752 2.013 2.013 0 0 0-.335-.608 1.41 1.41 0 0 0-.527-.41 1.65 1.65 0 0 0-.731-.151c-.3 0-.556.064-.766.191-.21.128-.383.303-.52.527-.131.223-.23.483-.293.779a4.507 4.507 0 0 0-.096.95Zm14.198 2.277V8.989c0-.264-.047-.492-.143-.683a1.006 1.006 0 0 0-.438-.445c-.191-.104-.433-.157-.724-.157a1.64 1.64 0 0 0-.698.137 1.127 1.127 0 0 0-.458.369.89.89 0 0 0-.164.526h-1.64c0-.291.07-.574.212-.847.14-.274.346-.518.615-.732.269-.214.59-.383.964-.506a4.01 4.01 0 0 1 1.257-.184c.556 0 1.049.093 1.477.28.433.187.773.47 1.019.848.25.373.376.843.376 1.408v3.288c0 .337.022.64.068.91.05.263.12.494.212.69V14h-1.689a2.913 2.913 0 0 1-.184-.677 5.514 5.514 0 0 1-.062-.806Zm.24-3.015.013 1.018h-1.182c-.306 0-.575.03-.807.09a1.657 1.657 0 0 0-.581.245c-.155.11-.271.242-.349.397a1.16 1.16 0 0 0-.116.526c0 .196.046.376.137.54.091.16.223.285.396.376.178.091.392.137.643.137.337 0 .631-.068.882-.205.255-.141.455-.312.601-.513.146-.205.224-.399.233-.58l.533.73a2.502 2.502 0 0 1-.28.602c-.133.214-.306.42-.52.615a2.584 2.584 0 0 1-1.77.656c-.483 0-.914-.095-1.292-.286a2.297 2.297 0 0 1-.89-.787 2.039 2.039 0 0 1-.32-1.128c0-.391.073-.738.218-1.039.15-.305.37-.56.657-.765.291-.205.647-.36 1.066-.465.42-.11.898-.164 1.436-.164h1.292Zm8.278-1.477v8.819h-1.648V6.604h1.518l.13 1.421Zm4.82 2.208v.144c0 .538-.065 1.037-.192 1.497a3.656 3.656 0 0 1-.554 1.196c-.241.338-.54.6-.895.786-.356.187-.766.28-1.23.28-.461 0-.865-.084-1.21-.252a2.48 2.48 0 0 1-.869-.732 3.969 3.969 0 0 1-.574-1.107 7.113 7.113 0 0 1-.301-1.408v-.554c.06-.542.16-1.034.3-1.477a4 4 0 0 1 .575-1.141c.237-.32.526-.565.868-.738.342-.174.74-.26 1.197-.26.464 0 .877.09 1.237.273.36.178.663.433.909.766.246.328.43.724.554 1.19.123.46.184.972.184 1.537Zm-1.648.144v-.144c0-.341-.032-.658-.096-.95a2.482 2.482 0 0 0-.3-.78 1.473 1.473 0 0 0-.527-.519 1.432 1.432 0 0 0-.759-.191c-.291 0-.542.05-.752.15-.21.096-.385.23-.526.404-.141.173-.25.376-.328.608a3.644 3.644 0 0 0-.164.745v1.326c.054.329.148.63.28.903.132.273.319.492.56.656.247.16.561.24.944.24.296 0 .549-.065.759-.192.21-.128.38-.303.512-.527.137-.227.237-.49.301-.786.064-.296.096-.61.096-.943ZM92.852 3.5V14H91.21V3.5h1.64Zm-.288 6.528-.533-.007a4.82 4.82 0 0 1 .212-1.415c.141-.432.337-.808.588-1.127a2.616 2.616 0 0 1 2.099-1.012c.364 0 .692.05.984.15.296.1.552.262.766.486.214.218.376.505.485.86.114.352.17.78.17 1.286V14h-1.653V9.235c0-.355-.053-.638-.158-.847a.911.911 0 0 0-.444-.451 1.625 1.625 0 0 0-.718-.144c-.296 0-.558.06-.786.178a1.64 1.64 0 0 0-.56.485c-.15.205-.265.442-.342.711a3.28 3.28 0 0 0-.11.861Zm6.126.356v-.157c0-.534.077-1.028.232-1.484.155-.46.378-.859.67-1.196a3.022 3.022 0 0 1 1.08-.793 3.513 3.513 0 0 1 1.449-.287c.542 0 1.025.096 1.449.287.429.187.791.451 1.087.793.296.337.522.736.677 1.196.155.456.232.95.232 1.484v.157c0 .533-.077 1.027-.232 1.483a3.555 3.555 0 0 1-.677 1.196 3.132 3.132 0 0 1-1.08.793c-.424.187-.904.28-1.442.28a3.603 3.603 0 0 1-1.456-.28 3.136 3.136 0 0 1-1.08-.793 3.55 3.55 0 0 1-.677-1.196 4.579 4.579 0 0 1-.233-1.483Zm1.647-.157v.157c0 .332.034.647.102.943.069.296.176.556.322.78.146.223.332.398.56.526.228.127.499.191.814.191.305 0 .569-.064.793-.191.228-.128.414-.303.56-.527.146-.223.253-.483.322-.779.072-.296.109-.61.109-.943v-.157c0-.329-.037-.638-.109-.93a2.39 2.39 0 0 0-.329-.786 1.56 1.56 0 0 0-1.36-.732c-.31 0-.579.066-.807.199a1.59 1.59 0 0 0-.553.533c-.146.228-.253.49-.322.786a4.085 4.085 0 0 0-.102.93Zm10.001-3.623v1.203h-4.17V6.604h4.17Zm-2.967-1.812h1.648v7.164c0 .228.031.403.095.526a.514.514 0 0 0 .281.24c.118.04.257.061.417.061.113 0 .223-.007.328-.02.104-.014.189-.028.253-.041l.006 1.258a4.175 4.175 0 0 1-.478.109 3.517 3.517 0 0 1-.615.048c-.379 0-.714-.066-1.005-.199a1.478 1.478 0 0 1-.684-.663c-.164-.305-.246-.71-.246-1.216V4.792Zm3.76 5.592v-.157c0-.534.077-1.028.232-1.484.155-.46.379-.859.67-1.196a3.02 3.02 0 0 1 1.08-.793 3.52 3.52 0 0 1 1.449-.287c.543 0 1.026.096 1.45.287.428.187.79.451 1.087.793.296.337.521.736.676 1.196.155.456.233.95.233 1.484v.157c0 .533-.078 1.027-.233 1.483-.155.456-.38.855-.676 1.196a3.15 3.15 0 0 1-1.08.793c-.424.187-.905.28-1.443.28a3.61 3.61 0 0 1-1.456-.28 3.132 3.132 0 0 1-1.08-.793 3.555 3.555 0 0 1-.677-1.196 4.579 4.579 0 0 1-.232-1.483Zm1.647-.157v.157c0 .332.034.647.103.943.068.296.175.556.321.78.146.223.333.398.561.526.228.127.499.191.813.191.306 0 .57-.064.793-.191a1.59 1.59 0 0 0 .561-.527c.146-.223.253-.483.321-.779a3.93 3.93 0 0 0 .109-.943v-.157c0-.329-.036-.638-.109-.93a2.389 2.389 0 0 0-.328-.786 1.57 1.57 0 0 0-.561-.533 1.542 1.542 0 0 0-.8-.199 1.58 1.58 0 0 0-.806.199 1.592 1.592 0 0 0-.554.533c-.146.228-.253.49-.321.786a4.029 4.029 0 0 0-.103.93Z"
-    />
-  </svg>
-);
-export default SVGInput;
